Cleanup the code, refactor existing libs for readability
