+++ runners/samza/src/main/java/org/apache/beam/runners/samza/SamzaRunner.java
@@ -29,6 +29,7 @@ import org.apache.beam.runners.core.construction.renderer.PipelineDotRenderer;
 import org.apache.beam.runners.fnexecution.provisioning.JobInfo;
 import org.apache.beam.runners.jobsubmission.PortablePipelineResult;
 import org.apache.beam.runners.samza.translation.ConfigBuilder;
+import org.apache.beam.runners.samza.translation.ConfigContext;
 import org.apache.beam.runners.samza.translation.PViewToIdMapper;
 import org.apache.beam.runners.samza.translation.PortableTranslationContext;
 import org.apache.beam.runners.samza.translation.SamzaPipelineTranslator;
@@ -128,27 +129,21 @@ public class SamzaRunner extends PipelineRunner<SamzaPipelineResult> {
       LOG.debug(
           "Pre-processed Beam pipeline in dot format:\n{}",
           PipelineDotRenderer.toDotString(pipeline));
-      LOG.debug(
-          "Pre-processed Beam pipeline in json format:\n{}",
-          PipelineJsonRenderer.toJsonString(pipeline, Collections.emptyMap()));
     }
 
     pipeline.replaceAll(SamzaTransformOverrides.getDefaultOverrides());
-
     final Map<PValue, String> idMap = PViewToIdMapper.buildIdMap(pipeline);
     final Set<String> nonUniqueStateIds = StateIdParser.scan(pipeline);
-    final Map<String, Map.Entry<String, String>> transformIOMap =
-        SamzaPipelineTranslator.buildTransformIOMap(pipeline, options, idMap, nonUniqueStateIds);
+    final ConfigContext configCtx = new ConfigContext(idMap, nonUniqueStateIds, options);
 
     final String dotGraph = PipelineDotRenderer.toDotString(pipeline);
     LOG.info("Beam pipeline DOT graph:\n{}", dotGraph);
 
-    final String jsonGraph = PipelineJsonRenderer.toJsonString(pipeline, transformIOMap);
+    final String jsonGraph = PipelineJsonRenderer.toJsonString(pipeline, configCtx);
     LOG.info("Beam pipeline JSON graph:\n{}", jsonGraph);
 
     final ConfigBuilder configBuilder = new ConfigBuilder(options);
-    SamzaPipelineTranslator.createConfig(
-        pipeline, options, idMap, nonUniqueStateIds, configBuilder);
+    SamzaPipelineTranslator.createConfig(pipeline, configCtx, configBuilder);
     configBuilder.put(BEAM_DOT_GRAPH, dotGraph);
     configBuilder.put(BEAM_JSON_GRAPH, jsonGraph);
