+++ runners/samza/src/main/java/org/apache/beam/runners/samza/translation/SamzaPipelineTranslator.java
@@ -20,25 +20,16 @@ package org.apache.beam.runners.samza.translation;
 import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkArgument;
 
 import com.google.auto.service.AutoService;
-import java.util.AbstractMap;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.ServiceLoader;
-import java.util.Set;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
 import org.apache.beam.runners.core.construction.PTransformTranslation;
 import org.apache.beam.runners.core.construction.TransformPayloadTranslatorRegistrar;
 import org.apache.beam.runners.core.construction.graph.ExecutableStage;
-import org.apache.beam.runners.samza.SamzaPipelineOptions;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.runners.TransformHierarchy;
 import org.apache.beam.sdk.transforms.Combine;
 import org.apache.beam.sdk.transforms.PTransform;
-import org.apache.beam.sdk.values.PCollection;
-import org.apache.beam.sdk.values.PValue;
-import org.apache.beam.sdk.values.TupleTag;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableMap;
 
 /** This class knows all the translators from a primitive BEAM transform to a Samza operator. */
@@ -48,7 +39,6 @@ import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.Immutabl
 })
 public class SamzaPipelineTranslator {
 
-  private static final String TRANSFORM_IO_MAP_DELIMITER = ",";
   private static final Map<String, TransformTranslator<?>> TRANSLATORS = loadTranslators();
 
   private static Map<String, TransformTranslator<?>> loadTranslators() {
@@ -83,12 +73,7 @@ public class SamzaPipelineTranslator {
   }
 
   public static void createConfig(
-      Pipeline pipeline,
-      SamzaPipelineOptions options,
-      Map<PValue, String> idMap,
-      Set<String> nonUniqueStateIds,
-      ConfigBuilder configBuilder) {
-    final ConfigContext ctx = new ConfigContext(idMap, nonUniqueStateIds, options);
+      Pipeline pipeline, ConfigContext ctx, ConfigBuilder configBuilder) {
 
     final TransformVisitorFn configFn =
         new TransformVisitorFn() {
@@ -114,48 +99,7 @@ public class SamzaPipelineTranslator {
     pipeline.traverseTopologically(visitor);
   }
 
-  /**
-   * Builds a map from PTransform to its input and output PValues. The map is serialized and stored
-   * in the job config.
-   */
-  public static Map<String, Map.Entry<String, String>> buildTransformIOMap(
-      Pipeline pipeline,
-      SamzaPipelineOptions options,
-      Map<PValue, String> idMap,
-      Set<String> nonUniqueStateIds) {
-    final ConfigContext ctx = new ConfigContext(idMap, nonUniqueStateIds, options);
-
-    final Map<String, Map.Entry<String, String>> pTransformToInputOutputMap = new HashMap<>();
-    final TransformVisitorFn configFn =
-        new TransformVisitorFn() {
-          @Override
-          public <T extends PTransform<?, ?>> void apply(
-              T transform,
-              TransformHierarchy.Node node,
-              Pipeline pipeline,
-              TransformTranslator<T> translator) {
-            ctx.setCurrentTransform(node.toAppliedPTransform(pipeline));
-            List<String> inputs = getIOPValueList(node.getInputs()).get();
-            List<String> outputs = getIOPValueList(node.getOutputs()).get();
-            pTransformToInputOutputMap.put(
-                node.getFullName(),
-                new AbstractMap.SimpleEntry<>(
-                    String.join(TRANSFORM_IO_MAP_DELIMITER, inputs),
-                    String.join(TRANSFORM_IO_MAP_DELIMITER, outputs)));
-            ctx.clearCurrentTransform();
-          }
-        };
-
-    final SamzaPipelineVisitor visitor = new SamzaPipelineVisitor(configFn);
-    pipeline.traverseTopologically(visitor);
-    return pTransformToInputOutputMap;
-  }
-
-  private static Supplier<List<String>> getIOPValueList(Map<TupleTag<?>, PCollection<?>> map) {
-    return () -> map.values().stream().map(pColl -> pColl.getName()).collect(Collectors.toList());
-  }
-
-  private interface TransformVisitorFn {
+  public interface TransformVisitorFn {
     <T extends PTransform<?, ?>> void apply(
         T transform,
         TransformHierarchy.Node node,
@@ -163,10 +107,10 @@ public class SamzaPipelineTranslator {
         TransformTranslator<T> translator);
   }
 
-  private static class SamzaPipelineVisitor extends Pipeline.PipelineVisitor.Defaults {
+  public static class SamzaPipelineVisitor extends Pipeline.PipelineVisitor.Defaults {
     private final TransformVisitorFn visitorFn;
 
-    private SamzaPipelineVisitor(TransformVisitorFn visitorFn) {
+    public SamzaPipelineVisitor(TransformVisitorFn visitorFn) {
       this.visitorFn = visitorFn;
     }
