+++ runners/samza/src/test/java/org/apache/beam/runners/samza/util/PipelineJsonRendererTest.java
@@ -24,12 +24,14 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.Collections;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 import org.apache.beam.runners.samza.SamzaExecutionEnvironment;
 import org.apache.beam.runners.samza.SamzaPipelineOptions;
 import org.apache.beam.runners.samza.SamzaRunner;
+import org.apache.beam.runners.samza.translation.ConfigBuilder;
+import org.apache.beam.runners.samza.translation.ConfigContext;
 import org.apache.beam.runners.samza.translation.PViewToIdMapper;
 import org.apache.beam.runners.samza.translation.SamzaPipelineTranslator;
 import org.apache.beam.runners.samza.translation.SamzaTransformOverrides;
@@ -37,6 +39,8 @@ import org.apache.beam.runners.samza.translation.StateIdParser;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.Create;
+import org.apache.beam.sdk.transforms.Filter;
+import org.apache.beam.sdk.transforms.Impulse;
 import org.apache.beam.sdk.transforms.Sum;
 import org.apache.beam.sdk.transforms.windowing.FixedWindows;
 import org.apache.beam.sdk.transforms.windowing.Window;
@@ -56,6 +60,9 @@ public class PipelineJsonRendererTest {
     options.setRunner(SamzaRunner.class);
 
     Pipeline p = Pipeline.create(options);
+    final Map<PValue, String> idMap = PViewToIdMapper.buildIdMap(p);
+    final Set<String> nonUniqueStateIds = StateIdParser.scan(p);
+    final ConfigContext ctx = new ConfigContext(idMap, nonUniqueStateIds, options);
 
     String jsonDag =
         "{  \"RootNode\": ["
@@ -66,8 +73,7 @@ public class PipelineJsonRendererTest {
     assertEquals(
         JsonParser.parseString(jsonDag),
         JsonParser.parseString(
-            PipelineJsonRenderer.toJsonString(p, Collections.emptyMap())
-                .replaceAll(System.lineSeparator(), "")));
+            PipelineJsonRenderer.toJsonString(p, ctx).replaceAll(System.lineSeparator(), "")));
   }
 
   @Test
@@ -87,17 +93,51 @@ public class PipelineJsonRendererTest {
 
     final Map<PValue, String> idMap = PViewToIdMapper.buildIdMap(p);
     final Set<String> nonUniqueStateIds = StateIdParser.scan(p);
+    final ConfigContext ctx = new ConfigContext(idMap, nonUniqueStateIds, options);
 
     String jsonDagFileName = "src/test/resources/ExpectedDag.json";
     String jsonDag =
         new String(Files.readAllBytes(Paths.get(jsonDagFileName)), StandardCharsets.UTF_8);
-
-    String renderedDag =
-        PipelineJsonRenderer.toJsonString(
-            p, SamzaPipelineTranslator.buildTransformIOMap(p, options, idMap, nonUniqueStateIds));
+    String renderedDag = PipelineJsonRenderer.toJsonString(p, ctx);
 
     assertEquals(
         JsonParser.parseString(jsonDag),
         JsonParser.parseString(renderedDag.replaceAll(System.lineSeparator(), "")));
+  }
+
+  @Test
+  public void testBeamTransformIOConfigGen() {
+    SamzaPipelineOptions options = PipelineOptionsFactory.create().as(SamzaPipelineOptions.class);
+    options.setJobName("TestEnvConfig");
+    options.setRunner(SamzaRunner.class);
+    options.setSamzaExecutionEnvironment(SamzaExecutionEnvironment.LOCAL);
+
+    Pipeline pipeline = Pipeline.create(options);
+    pipeline.apply(Impulse.create()).apply(Filter.by(Objects::nonNull));
+    pipeline.replaceAll(SamzaTransformOverrides.getDefaultOverrides());
+
+    final Map<PValue, String> idMap = PViewToIdMapper.buildIdMap(pipeline);
+    final Set<String> nonUniqueStateIds = StateIdParser.scan(pipeline);
+    final ConfigContext configCtx = new ConfigContext(idMap, nonUniqueStateIds, options);
+
+    final ConfigBuilder configBuilder = new ConfigBuilder(options);
+    SamzaPipelineTranslator.createConfig(pipeline, configCtx, configBuilder);
+    final Map<String, Map.Entry<String, String>> transformInputOutput =
+        PipelineJsonRenderer.buildTransformIOMap(pipeline, configCtx);
+
+    assertEquals(2, transformInputOutput.size());
+    assertEquals("", transformInputOutput.get("Impulse").getKey()); // no input to impulse
+    assertEquals(
+        "Impulse.out",
+        transformInputOutput.get("Impulse").getValue()); // PValue for to Impulse.output
+
+    // Input to Filter is PValue Output from Impulse
+    assertEquals(
+        "Impulse.out",
+        transformInputOutput.get("Filter/ParDo(Anonymous)/ParMultiDo(Anonymous)").getKey());
+    // output PValue of filter
+    assertEquals(
+        "Filter/ParDo(Anonymous)/ParMultiDo(Anonymous).output",
+        transformInputOutput.get("Filter/ParDo(Anonymous)/ParMultiDo(Anonymous)").getValue());
   }
 }
